✏️ Test: #12 플레이어 스킬 데이터 파싱 테스트
- `PlayerSkillData` 파싱 로직 테스트
- JSON 데이터 수정: #12 에서 발생했던 문제를 해결하기 위해 임시로 `CastTime`을 수정
- `PlayerSkillData` 어드레서블에 추가

+++ Project-C/Assets/@Scripts/Character/Creature.cs
@@ -36,12 +36,16 @@ public class Creature : BaseObject
         }
     }
 
+    #region Stat
     public int Hp { get; protected set; }
     public int MaxHp { get; protected set; }
-    public int Atk { get; protected set; }
+    public float Atk { get; protected set; }
     public float MoveSpeed { get; protected set; }
     public float JumpForce { get; protected set; }
     public float DoubleJumpForce { get; protected set; }    // 이단 점프할 때 추가적인 점프 힘
+    #endregion
+
+    //public List<SkillBase> Skills { get; protected set; }   // TODO
 
     public override bool Init()
     {
